fix(ui): paging works after clicking in trace table
